Move security checks to a separate module
